pyproject.toml
[tool.poetry]
name = "amocrm-api"
version = "1.0.0"
description = "Amocrm api connector python"
authors = ["Alexander Dr. <sasna142@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
loguru = "^0.5.3"
pydantic = {extras = ["email"], version = "^1.8.2"}
pandas = "^1.2.4"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
